pyproject.toml
[tool.poetry]
name = "td2"
version = "0.1.0"
description = ""
authors = ["mdeboute <martin.deboute@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
python-chess = "^1.999"
numpy = "^1.21.4"
scipy = "^1.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
